// src/lib/api-client/generated/api.schemas.ts
export type WbpItemJenisKelamin =
  (typeof WbpItemJenisKelamin)[keyof typeof WbpItemJenisKelamin];

export const WbpItemJenisKelamin = {
  L: "L",
  P: "P",
} as const;

export type WbpItemJenisLayanan =
  (typeof WbpItemJenisLayanan)[keyof typeof WbpItemJenisLayanan];

export const WbpItemJenisLayanan = {
  PB: "PB",
  CB: "CB",
  CMB: "CMB",
  ASIMILASI: "ASIMILASI",
} as const;

export type WbpItemTahapSaatIni =
  (typeof WbpItemTahapSaatIni)[keyof typeof WbpItemTahapSaatIni];

export const WbpItemTahapSaatIni = {
  verifikasi_rutan: "verifikasi_rutan",
  pengusulan_litmas: "pengusulan_litmas",
  sidang_tpp_upt: "sidang_tpp_upt",
  upload_sdp: "upload_sdp",
  verifikasi_kanwil: "verifikasi_kanwil",
  proses_ditjen_pas: "proses_ditjen_pas",
  sk_terbit: "sk_terbit",
  turun_sk: "turun_sk",
} as const;

export type WbpItemStatus = (typeof WbpItemStatus)[keyof typeof WbpItemStatus];

export const WbpItemStatus = {
  aktif: "aktif",
  selesai: "selesai",
  ditolak: "ditolak",
} as const;

export interface WbpItem {
  id: string;
  kodeTracking?: string;
  nama: string;
  nomorRegistrasi: string;
  jenisKelamin?: WbpItemJenisKelamin;
  tempatLahir?: string;
  tanggalLahir?: string;
  nomorHpKeluarga?: string;
  namaKontakKeluarga?: string;
  perkara?: string;
  alamat?: string;
  tanggalPelaksanaan?: string;
  jenisLayanan: WbpItemJenisLayanan;
  tahapSaatIni: WbpItemTahapSaatIni;
  status: WbpItemStatus;
  catatan?: string;
  createdAt: string;
  updatedAt: string;
}